backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
websockets==14.2
docker==7.1.0
pyyaml==6.0.2
jinja2==3.1.5
pydantic==2.11.1
